+++ src/containerapp/HISTORY.rst
@@ -6,6 +6,7 @@ Upcoming
 +++++++
 * Fix the 'TypeError: 'NoneType' object does not support item assignment' error obtained while running the CLI command 'az containerapp dapr enable'
 * Update API version to stable "2022-10-01".
+* Updating quickstart image
 
 0.3.21
 ++++++